Move MetricsHandler to new package and add 'V1' to its name.
