+++ metrics-proxy/src/test/java/ai/vespa/metricsproxy/http/metrics/MetricsV1HandlerTest.java
@@ -1,6 +1,7 @@
 // Copyright 2020 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.metricsproxy.http;
+package ai.vespa.metricsproxy.http.metrics;
 
+import ai.vespa.metricsproxy.http.HttpHandlerTestBase;
 import ai.vespa.metricsproxy.metric.model.json.GenericJsonModel;
 import ai.vespa.metricsproxy.metric.model.json.GenericMetrics;
 import ai.vespa.metricsproxy.metric.model.json.GenericService;
@@ -17,8 +18,8 @@ import java.io.IOException;
 import java.util.concurrent.Executors;
 
 import static ai.vespa.metricsproxy.core.VespaMetrics.INSTANCE_DIMENSION_ID;
-import static ai.vespa.metricsproxy.http.MetricsHandler.V1_PATH;
-import static ai.vespa.metricsproxy.http.MetricsHandler.VALUES_PATH;
+import static ai.vespa.metricsproxy.http.metrics.MetricsV1Handler.V1_PATH;
+import static ai.vespa.metricsproxy.http.metrics.MetricsV1Handler.VALUES_PATH;
 import static ai.vespa.metricsproxy.metric.model.StatusCode.DOWN;
 import static ai.vespa.metricsproxy.metric.model.json.JacksonUtil.createObjectMapper;
 import static ai.vespa.metricsproxy.service.DummyService.METRIC_1;
@@ -33,17 +34,17 @@ import static org.junit.Assert.fail;
  * @author gjoranv
  */
 @SuppressWarnings("UnstableApiUsage")
-public class MetricsHandlerTest extends HttpHandlerTestBase {
+public class MetricsV1HandlerTest extends HttpHandlerTestBase {
 
     private static final String V1_URI = URI_BASE + V1_PATH;
     private static final String VALUES_URI = URI_BASE + VALUES_PATH;
 
     @BeforeClass
     public static void setup() {
-        MetricsHandler handler = new MetricsHandler(Executors.newSingleThreadExecutor(),
-                                                    getMetricsManager(),
-                                                    vespaServices,
-                                                    getMetricsConsumers());
+        var handler = new MetricsV1Handler(Executors.newSingleThreadExecutor(),
+                                                        getMetricsManager(),
+                                                        vespaServices,
+                                                        getMetricsConsumers());
         testDriver = new RequestHandlerTestDriver(handler);
     }
